refactor(api): move exception handlers into exceptions packages
Relocate the exception→HTTP mechanism into dedicated packages:
core/exception_handlers.py -> core/exceptions/handlers.py, and the lib facade plus
domain classes into lib/exceptions/ (handlers.py + __init__.py). Update imports,
docstrings, CLAUDE.md, the plugin guide, and move the test to
tests/core/exceptions/test_handlers.py.
