+++ sparkth/lib/exceptions/__init__.py
